pyproject.toml
[project]
name = "kabus-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.28",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
    "xgboost>=3.2.0",
]

[tool.uv]
python-version = "3.12"
